Calculating the total weight of widgets and gizmos that are ordered
A program that reads the number of widgets and the number of gizmos in an order from the user. Then the program should compute and display the total weight of the order. Considering an online retailer sells two products: widgets and gizmos. Each widget weighs 75 grams and each gizmo weighs 112 grams.